Move InvoicePaid hook to includes/hooks/ for guaranteed WHMCS autoload
